bs_bos_core_base: t_ulong should be unsigned long long on Windows

+++ bs_bos_core_base/include/conf.h
@@ -44,7 +44,7 @@ typedef double        t_double;
 
 typedef unsigned long               t_uint;
 #ifdef _WIN32
-typedef unsigned long          t_ulong;
+typedef unsigned long long          t_ulong;
 #else
 typedef unsigned long               t_ulong;
 #endif
